refactor: handle login in appnavigation
