--- .expo/packager-info.json
@@ -1,9 +1,0 @@
-{
-  "devToolsPort": 19003,
-  "expoServerPort": 19000,
-  "packagerPort": 19004,
-  "packagerPid": 10035,
-  "expoServerNgrokUrl": "https://vb-c3y.dridridridri.app.exp.direct",
-  "packagerNgrokUrl": "https://packager.vb-c3y.dridridridri.app.exp.direct",
-  "ngrokPid": 10052
-}
