+++ components/stacks/ChoiseTeam.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { Text, View } from "react-native";
