+++ store/reducers/AuthReducer.js
@@ -3,7 +3,6 @@ import { LOGIN, LOGOUT } from "../actions/AuthActions"
 export default function AuthReducer(prevState, action) {
   switch (action.type) {
     case LOGIN:
-      console.log('HEEEEY')
       return {
         ...prevState,
         user: action.payload
